mosaic_backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==23.7.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.6
configparser==5.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
cryptography==39.0.1
defusedxml==0.7.1
Django==5.0.1
django-cors-headers==3.13.0
django-extensions==3.2.3
django-filter==23.5
django-mdeditor==0.1.20
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-templated-mail==1.1.1
django-timezone-utils==0.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docopt==0.6.2
drf-spectacular==0.27.1
drf-yasg==1.21.5
drf_api_logger==1.1.12
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.3.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.1
phonenumbers==8.13.5
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==3.1.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pydot==1.4.2
pydub==0.25.1
pyflakes==3.0.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0.1
referencing==0.32.1
requests==2.28.2
requests-oauthlib==1.3.1
rpds-py==0.17.1
ruamel.yaml==0.17.21
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sounddevice==0.4.6
sqlparse==0.4.3
testfixtures==6.18.5
tqdm==4.65.0
typing_extensions==4.2.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
yarg==0.1.9
zipp==3.8.0
docutils==0.20.1
django-anymail==10.3